Needed to unsubscribe before getTasks was called Moved task component to be under todo module
