+++ src/app/app.module.ts
@@ -34,12 +34,12 @@ import { LoginComponent } from './login/login.component';
 import { NavMenuComponent } from './navigation/nav-menu/nav-menu.component';
 import { NavigationComponent } from './navigation/navigation.component';
 import { RegisterComponent } from './register/register.component';
-import { TaskComponent } from './task/task.component';
 import { AddTaskFieldComponent } from './todo/add-task-field/add-task-field.component';
 import { DetailsComponent } from './todo/details/details.component';
 import { EditContentComponent } from './todo/edit-content/edit-content.component';
 import { FilterComponent } from './todo/filter/filter.component';
 import { TaskListComponent } from './todo/task-list/task-list.component';
+import { TaskComponent } from './todo/task/task.component';
 import { TodoComponent } from './todo/todo.component';
 
 @NgModule({
